+++ comics/comics/commitstrip.py
@@ -6,8 +6,7 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://www.commitstrip.com/en/'
     start_date = '2012-02-22'
-    rights = 'CommitStrip/Etienne Issartial'
-
+    rights = 'Etienne Issartial'
 
 class Crawler(CrawlerBase):
     history_capable_days = 30
@@ -19,5 +18,4 @@ class Crawler(CrawlerBase):
         for entry in feed.for_date(pub_date):
             url = entry.content0.src('img[src*="/uploads/"]')
             title = entry.title
-
             return CrawlerImage(url, title)