+++ docs/changes.rst
@@ -10,6 +10,7 @@ v2.3.2 (UNRELEASED)
 
 - Dependencies with new maximum versions:
 
+  - django_compressor >= 1.1, < 1.5
   - Pillow >= 1.7, < 2.6
 
 **Crawlers**