+++ docs/changes.rst
@@ -22,6 +22,7 @@ v2.3.2 (UNRELEASED)
 - New: ``blasternation``
 - New: ``cardboardcrack``
 - New: ``lunche24``
+- New: ``q2qcomics``
 - Update: ``asofterworld`` after site change.
 - Update: ``doghouse`` after feed change.
 - Update: ``joyoftech`` after site change.